requirements.txt
absl-py==2.1.0
ansicon==1.89.0
astunparse==1.6.3
async-timeout==5.0.1
blessed==1.20.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
Deprecated==1.2.18
filelock==3.17.0
Flask==3.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.2.0
gast==0.6.0
gevent==24.11.1
google-pasta==0.2.0
gpustat==1.1.1
GPUtil==1.4.0
greenlet==3.1.1
grpcio==1.70.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jinxed==1.3.0
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
limits==4.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
nvidia-ml-py==12.570.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
ordered-set==4.1.0
packaging==24.2
pillow==11.1.0
prometheus_client==0.21.1
protobuf==5.29.3
psutil==7.0.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-magic-bin==0.4.14
redis==5.2.1
requests==2.31.0
rich==13.9.4
six==1.17.0
SQLAlchemy==1.4.51
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
torch==2.6.0
torchaudio
torchvision==0.21.0
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zope.event==5.0
zope.interface==7.2
flask
flask-jwt-extended
flask-wtf
werkzeug
sqlalchemy
python-dotenv
gunicorn